backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
asyncpg==0.29.0
redis==5.0.7
httpx==0.27.0
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
faiss-cpu==1.8.0
sentence-transformers==3.0.1
numpy==1.26.4
boto3==1.34.144
aiofiles==24.1.0
